refactor: backport some changes from merkle integration
