+++ src/rpc/serviceDefinitions/syncDefinition.ts
@@ -1,6 +1,6 @@
 import { GetAllMessagesByFidRequest, MessagesResponse } from '~/flatbuffers/generated/rpc_generated';
 import { toByteBuffer } from '~/flatbuffers/utils/bytes';
-import { defaultMethod } from '~/rpc/server';
+import { defaultMethod } from '~/rpc/client';
 
 const defaultSyncMethod = () => {
   return {
